Fix landing login button by restoring modal
